// src/app/interfaces/passings-data-obj.ts
export interface PassingsDataObj {
    absolute_magnitude_h: number;
    close_approach_data: {}[];
    designation: string;
    estimated_diameter: {};
    id: string;
    is_potentially_hazardous_asteroid: boolean;
    is_sentry_object: boolean;
    links: {};
    name: string;
    nasa_jpl_url: string;
    neo_reference_id: string;
    orbital_data: {};
}